Separate the class and finalize test code
